requirements.txt
boto3==1.35.0
bcrypt==4.1.2
PyJWT==2.8.0